# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/nodes/storagenodes": {
            "get": {
                "description": "list StorageNode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "摘要 获取存储节点列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "Name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StorageNodeList"
                        }
                    }
                }
            }
        },
        "/nodes/storagenodes/:storagenodename": {
            "get": {
                "description": "get StorageNode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "摘要 获取指定存储节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "Name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StorageNode"
                        }
                    }
                }
            }
        },
        "/nodes/storagenodes/:storagenodename/disks": {
            "get": {
                "description": "list StorageNodeDisks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "摘要 获取指定存储节点磁盘列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "nodeName",
                        "name": "NodeName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.LocalDiskListByNode"
                        }
                    }
                }
            }
        },
        "/nodes/storagenodes/:storagenodename/migrates": {
            "get": {
                "description": "get StorageNodeMigrate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Node"
                ],
                "summary": "摘要 获取指定节点数据卷迁移任务列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "nodeName",
                        "name": "NodeName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.VolumeOperationList"
                        }
                    }
                }
            }
        },
...
        "/volumes/volumeoperations": {
            "get": {
                "description": "list VolumeOperations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Volume"
                ],
                "summary": "摘要 获取数据卷操作记录列表信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "Name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/api.VolumeOperationList"
                        }
                    }
                }
            }
        },
        "/volumes/volumeoperations/:volumename": {
            "get": {
                "description": "get VolumeOperation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Volume"
                ],
                "summary": "摘要 获取指定数据卷操作记录信息（目前仅包含迁移运维操作）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "volumeName",
                        "name": "VolumeName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/api.VolumeOperation"
                        }
                    }
                }
            }
        },
        "/volumes/volumeoperations/yamls/:operationname": {
            "get": {
                "description": "get VolumeOperationYamlGet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Volume"
                ],
                "summary": "摘要 获取数据卷操作记录yaml信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "Name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/api.LocalVolumeMigrate"
                        }
                    }
                }
            }
        },
        "/volumes/volumereplicas/:volumename": {
            "get": {
                "description": "list volumes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Volume"
                ],
                "summary": "摘要 获取指定数据卷的副本列表信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "volumeName",
                        "name": "VolumeName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/api.VolumeReplicaList"
                        }
                    }
                }
            }
        },
        "/volumes/volumereplicas/yamls/:volumereplicaname": {
            "get": {
                "description": "get VolumeReplicaYaml",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Volume"
                ],
                "summary": "摘要 查看指定数据卷副本yaml信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "Name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/api.LocalVolumeReplica"
                        }
                    }
                }
            }
        },
        "/volumes/volumes": {
            "get": {
                "description": "list Volume",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Volume"
                ],
                "summary": "摘要 获取数据卷列表信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "Name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/api.VolumeList"
                        }
                    }
                }
            }
        },
        "/volumes/volumes/:volumename": {
            "get": {
                "description": "get Volume",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Volume"
                ],
                "summary": "摘要 获取指定数据卷基本信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "Name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Volume"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.LocalDisk": {
            "type": "object",
            "properties": {
                "availableCapacityBytes": {
                    "description": "AvailableCapacityBytes 可用容量",
                    "type": "integer"
                },
                "devPath": {
                    "description": "e.g. /dev/sdb 磁盘路径",
                    "type": "string"
                },
                "hasRaid": {
                    "description": "HasRAID 是否Raid",
                    "type": "boolean"
                },
                "localStoragePooLName": {
                    "description": "LocalStoragePooLName 存储池名称",
                    "type": "string"
                },
                "state": {
                    "description": "Possible state: Claimed, UnClaimed, Inuse, Released, Reserved 状态",
                    "type": "string"
                },
                "totalCapacityBytes": {
                    "description": "TotalCapacityBytes 总容量",
                    "type": "integer"
                },
                "type": {
                    "description": "Supported: HDD, SSD, NVMe, RAM 磁盘类型",
                    "type": "string"
                }
            }
        },
        "api.LocalDiskListByNode": {
            "type": "object",
            "properties": {
                "localDisks": {
                    "description": "localDisks 节点磁盘列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.LocalDisk"
                    }
                },
                "nodeName": {
                    "description": "nodeName 节点名称",
                    "type": "string"
                },
                "page": {
                    "description": "page",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "pageSize",
                    "type": "integer"
                }
            }
        },
        "api.LocalVolumeMigrate": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        },
                        "namespace": {
                            "type": "string"
                        }
                    }
                },
                "spec": {
                    "$ref": "#/definitions/api.LocalVolumeMigrateSpec"
                },
                "status": {
                    "$ref": "#/definitions/api.LocalVolumeMigrateStatus"
                }
            }
        },
        "api.LocalVolumeMigrateSpec": {
            "type": "object",
            "properties": {
                "abort": {
                    "description": "abort",
                    "type": "boolean"
                },
                "migrateAllVols": {
                    "description": "migrateAllVols",
                    "type": "boolean"
                },
                "sourceNode": {
                    "description": "sourceNode",
                    "type": "string"
                },
                "targetNodesSuggested": {
                    "description": "targetNodesSuggested",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "volumeName": {
                    "description": "volumeName",
                    "type": "string"
                }
            }
        },
        "api.LocalVolumeMigrateStatus": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "error message to describe some states",
                    "type": "string"
                },
                "originalReplicaNumber": {
                    "description": "record the volume's replica number, it will be set internally",
                    "type": "integer"
                },
                "state": {
                    "description": "State of the operation, e.g. submitted, started, completed, abort, ...",
                    "type": "string"
                },
                "targetNode": {
                    "description": "record the node where the specified replica is migrated to",
                    "type": "string"
                }
            }
        },
        "api.LocalVolumeReplica": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        },
                        "namespace": {
                            "type": "string"
                        }
                    }
                },
                "spec": {
                    "$ref": "#/definitions/api.LocalVolumeReplicaSpec"
                },
                "status": {
                    "$ref": "#/definitions/api.LocalVolumeReplicaStatus"
                }
            }
        },
...
        "api.StorageNode": {
            "type": "object",
            "properties": {
                "driverStatus": {
                    "description": "driver status 驱动状态",
                    "type": "string"
                },
                "freeCapacityBytes": {
                    "description": "freeCapacityBytes LSN可分配存储容量",
                    "type": "integer"
                },
                "ip": {
                    "description": "ip 节点IP",
                    "type": "string"
                },
                "isRaid": {
                    "description": "IsRAID 是否Raid",
                    "type": "boolean"
                },
                "name": {
                    "description": "name 节点名字",
                    "type": "string"
                },
                "nodeState": {
                    "description": "node state 节点状态",
                    "type": "string"
                },
                "totalDiskCount": {
                    "description": "totalDiskCount 总磁盘数",
                    "type": "integer"
                },
                "totalHDDCapacityBytes": {
                    "description": "totalHDDCapacityBytes HDD存储总容量",
                    "type": "integer"
                },
                "totalSSDCapacityBytes": {
                    "description": "totalSSDCapacityBytes SSD存储总容量",
                    "type": "integer"
                },
                "usedDiskCount": {
                    "description": "usedDiskCount 已纳管磁盘数",
                    "type": "integer"
                },
                "usedHDDCapacityBytes": {
                    "description": "usedHDDCapacityBytes HDD已经使用存储量",
                    "type": "integer"
                },
                "usedSSDCapacityBytes": {
                    "description": "usedSSDCapacityBytes SSD已经使用存储量",
                    "type": "integer"
                }
            }
        },
        "api.StorageNodeList": {
            "type": "object",
            "properties": {
                "page": {
                    "description": "page",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "pageSize",
                    "type": "integer"
                },
                "storageNodes": {
                    "description": "StorageNodes",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.StorageNode"
                    }
                }
            }
        },
...
        "api.Volume": {
            "type": "object",
            "properties": {
                "convertible": {
                    "description": "Convertible 转换高可用模式",
                    "type": "boolean"
                },
                "createTime": {
                    "description": "createTime 创建时间",
                    "type": "string"
                },
                "name": {
                    "description": "local volume name 名称",
                    "type": "string"
                },
                "pvcName": {
                    "description": "PersistentVolumeClaimName is the name of the associated PVC 绑定PVC",
                    "type": "string"
                },
                "pvcNamespace": {
                    "description": "PersistentVolumeClaimNamespace is the namespace of the associated PVC 命名空间",
                    "type": "string"
                },
                "replicaNumber": {
                    "description": "replica number 副本数",
                    "type": "integer"
                },
                "requiredCapacityBytes": {
                    "description": "size 容量",
                    "type": "integer"
                },
                "state": {
                    "description": "local volume state 状态",
                    "type": "string"
                },
                "volumegroup": {
                    "description": "VolumeGroup is the group name of the local volumes. It is designed for the scheduling and allocating. 磁盘组",
                    "type": "string"
                }
            }
        },
        "api.VolumeList": {
            "type": "object",
            "properties": {
                "page": {
                    "description": "page",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "pageSize",
                    "type": "integer"
                },
                "volumes": {
                    "description": "volumes",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Volume"
                    }
                }
            }
        },
        "api.VolumeMigrateOperation": {
            "type": "object",
            "properties": {
                "VolumeName": {
                    "description": "VolumeName 迁移卷名称",
                    "type": "string"
                },
                "endTime": {
                    "description": "EndTime 迁移结束时间",
                    "type": "string"
                },
                "name": {
                    "description": "VolumeMigrateName 迁移CRD名称",
                    "type": "string"
                },
                "sourceNode": {
                    "description": "SourceNode 迁移源节点",
                    "type": "string"
                },
                "startTime": {
                    "description": "StartTime 迁移开始时间",
                    "type": "string"
                },
                "state": {
                    "description": "State 迁移状态",
                    "type": "string"
                },
                "targetNode": {
                    "description": "TargetNode 迁移目的节点",
                    "type": "string"
                }
            }
        },
        "api.VolumeOperation": {
            "type": "object",
            "properties": {
                "VolumeMigrateOperations": {
                    "description": "VolumeMigrateOperations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.VolumeMigrateOperation"
                    }
                },
                "page": {
                    "description": "page",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "pageSize",
                    "type": "integer"
                },
                "volumeName": {
                    "description": "volume name",
                    "type": "string"
                }
            }
        },
        "api.VolumeOperationList": {
            "type": "object",
            "properties": {
                "VolumeOperations": {
                    "description": "VolumeOperations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.VolumeOperation"
                    }
                },
                "nodeName": {
                    "description": "node name",
                    "type": "string"
                }
            }
        },
        "api.VolumeReplica": {
            "type": "object",
            "properties": {
                "devicePath": {
                    "description": "DevicePath is a link path of the StoragePath of the volume replica,\ne.g. /dev/LocalStorage_PoolHDD/pvc-fbf3ffc3-66db-4dae-9032-bda3c61b8f85",
                    "type": "string"
                },
                "name": {
                    "description": "replica name",
                    "type": "string"
                },
                "nodeName": {
                    "description": "NodeName is the assigned node where the volume replica is located 节点",
                    "type": "string"
                },
                "requiredCapacityBytes": {
                    "description": "RequiredCapacityBytes 容量",
                    "type": "integer"
                },
                "state": {
                    "description": "replica state",
                    "type": "string"
                },
                "storagePath": {
                    "description": "StoragePath is a real path of the volume replica, like /dev/sdg.",
                    "type": "string"
                },
                "synced": {
                    "description": "Synced is the sync state of the volume replica, which is important in HA volume 同步状态",
                    "type": "boolean"
                }
            }
        },
        "api.VolumeReplicaList": {
            "type": "object",
            "properties": {
                "volumeName": {
                    "description": "volume name",
                    "type": "string"
                },
                "volumeReplicas": {
                    "description": "VolumeReplicas",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.VolumeReplica"
                    }
                }
            }
        }
    }
}